LayoutTests/userscripts/script-not-run-for-fragments-expected.txt

CONSOLE MESSAGE: line 1: injected
CONSOLE MESSAGE: line 1: injected
This test should not infinitely inject user scripts. 
